requirements.txt
bcrypt==4.3.0
cffi==1.17.1
cryptography==44.0.2
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.0
